requirements.txt
fpdf2
langgraph
langchain
langchain-community
langchain-ollama
pydantic
pytest
pypdf
duckduckgo-search
